Fix checking availability

+++ app/models/archive.rb
@@ -62,8 +62,10 @@ class Archive < ApplicationRecord
 
   def check_availability
     response = Faraday.head(url)
-    version.update_columns(yanked: true)
-    remove_pin if response.status == 404
-    self.destroy
+    if response.status == 404
+      version.update_columns(yanked: true)
+      remove_pin
+      self.destroy
+    end
   end
 end
